fix: keep scroll at top when navigating to /movie-database
Next.js's app router scroll-into-view targets the first DOM element of
the new page segment. With HeroSection and CuratedMediaRows living in
the (list) layout, the page boundary started at <Filters> below the
fold, so navigating from the home card landed the user inside the
discover grid.

Move HeroSection, CuratedMediaRows, and the InlineChatSwitch wrapper
into the page itself so the scroll target is the hero. The (list)
layout collapses into the page since it was the only consumer.
